+++ CHANGELOG.md
@@ -13,6 +13,7 @@
 - Removed `mixin` directive from operation string sent to server.
 - Fixed `ShorterResultsPlugin` that generated faulty code for discriminated unions.
 - Changed generator to ignore unused fragments which should be unpacked in queries.
+- Changed type hints for parse and serialize methods of scalars to `typing.Any`.
 
 
 ## 0.6.0 (2023-04-18)